# mbia-specs/technical/api/openapi.yaml
openapi: 3.1.0
info:
  title: Mbia API
  version: 0.2.0
  summary: REST contract for the Mbia commercial MVP
  description: |
    Contract-first API for Mbia, a private collaborative family heritage platform.

    Product specifications are authoritative for business behavior. This OpenAPI
    document is the authoritative HTTP contract derived from those specifications.
servers:
  - url: https://api.mbia.example.com/api/v1
    description: Production placeholder
  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /families/{familyId}/memories:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
    get:
      tags: [Memories]
      operationId: listFamilyMemories
      summary: List active family memories
      description: |
        ACTIVE Memories of the Family, most recently added first (`createdAt` descending,
        then `id`), for any ACTIVE member (OQ-034).
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/MemoryType'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Family memories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /families/{familyId}/memories/photos:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
    post:
      tags: [Memories]
      operationId: createPhotoMemory
      summary: Create a photo memory from an already uploaded READY media asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePhotoMemoryRequest'
      responses:
        '201':
          description: Photo memory created
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/memories/stories:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
    post:
      tags: [Memories]
      operationId: createStoryMemory
      summary: Create a written family story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoryMemoryRequest'
      responses:
        '201':
          description: Story created
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /families/{familyId}/memories/{memoryId}:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - $ref: '#/components/parameters/MemoryId'
    get:
      tags: [Memories]
      operationId: getMemory
      summary: Get a memory
      description: |
        Any ACTIVE member. An unknown, other-Family or ARCHIVED Memory returns 404
        `MEMORY_NOT_FOUND` (OQ-037).
      responses:
        '200':
          description: Memory
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Memories]
      operationId: updateMemory
      summary: Update a memory
      description: |
        The creator may edit their own memory; ADMIN may edit any family memory. Both need
        a role that can write: a VIEWER gets 403 `PERMISSION_DENIED` (OQ-041).
        A field irrelevant to the Memory type is rejected with 400 `VALIDATION_FAILED`.
        When the set of related Persons changes, the new set must include at least one ACTIVE
        Person (400 `VALIDATION_FAILED`, OQ-043); a Person newly added must be ACTIVE (409
        `PERSON_NOT_ACTIVE`), while an archived Person already on the Memory may stay (OQ-035). An unknown, other-Family or ARCHIVED Memory returns
        404 `MEMORY_NOT_FOUND` (OQ-037).
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemoryRequest'
      responses:
        '200':
          description: Updated memory
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Memories]
      operationId: archiveMemory
      summary: Archive a memory
      description: |
        The creator may archive their own memory; ADMIN may archive any family memory, with
        a role that can write (OQ-041). An archived Memory is no longer listed nor returned
        (404 `MEMORY_NOT_FOUND`).
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Memory archived
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/persons/{personId}/memories:
    parameters:
      - $ref: '#/components/parameters/FamilyId'
      - $ref: '#/components/parameters/PersonId'
    get:
      tags: [Memories]
      operationId: listPersonMemories
      summary: List memories associated with a person
      description: |
        ACTIVE Memories linked to the Person, most recently added first (`createdAt`
        descending, then `id`), for any ACTIVE member (OQ-034).
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Person memories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token validated by Spring Security.
  headers:
    ETag:
      description: Resource version encoded as an HTTP ETag, for example `"3"`.
      schema:
        type: string
        example: '"3"'
  parameters:
    FamilyId:
      name: familyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PersonId:
      name: personId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    MemoryId:
      name: memoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Expected resource ETag/version. Stale values return 409 CONCURRENT_MODIFICATION.
      schema:
        type: string
        example: '"3"'
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    Forbidden:
      description: Authenticated user is not authorized for this family/action
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found in the accessible scope
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Domain or optimistic-concurrency conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    DatePrecision:
      type: string
      enum: [EXACT, YEAR_ONLY, UNKNOWN]
    PartialDate:
      type: object
      required: [precision]
      description: |
        EXACT requires `date`; YEAR_ONLY requires `year`; UNKNOWN requires neither.
        The application must reject inconsistent combinations.
      properties:
        precision:
          $ref: '#/components/schemas/DatePrecision'
        date:
          type: [string, 'null']
          format: date
        year:
          type: [integer, 'null']
          minimum: 1
          maximum: 9999
    PersonStatus:
      type: string
      enum: [ACTIVE, ARCHIVED, MERGED]
...
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    MemoryType:
      type: string
      enum: [PHOTO, STORY]
    MemoryStatus:
      type: string
      enum: [ACTIVE, ARCHIVED]
    RelatedPersonReference:
      type: object
      required: [id, displayName, status]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        status:
          description: |
            An archived Person stays on the Memory and is shown as archived (OQ-035).
          allOf:
            - $ref: '#/components/schemas/PersonStatus'
        profilePictureUrl:
          type: [string, 'null']
          format: uri
    MemoryResponse:
      type: object
      required:
        - id
        - familyId
        - type
        - status
        - relatedPersons
        - createdBy
        - createdAt
        - updatedAt
        - version
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/MemoryType'
        status:
          $ref: '#/components/schemas/MemoryStatus'
        title:
          type: [string, 'null']
        content:
          type: [string, 'null']
        caption:
          type: [string, 'null']
        media:
          oneOf:
            - $ref: '#/components/schemas/MediaAssetResponse'
            - type: 'null'
        takenAt:
          oneOf:
            - $ref: '#/components/schemas/PartialDate'
            - type: 'null'
        relatedPersons:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RelatedPersonReference'
        createdBy:
          $ref: '#/components/schemas/ActivityActor'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    CreatePhotoMemoryRequest:
      type: object
      required: [mediaAssetId, relatedPersonIds]
      properties:
        mediaAssetId:
          type: string
          format: uuid
        caption:
          type: [string, 'null']
          maxLength: 5000
        takenAt:
          oneOf:
            - $ref: '#/components/schemas/PartialDate'
            - type: 'null'
        relatedPersonIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: uuid
    CreateStoryMemoryRequest:
      type: object
      required: [title, content, relatedPersonIds]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 250
        content:
          type: string
          minLength: 1
          maxLength: 50000
        relatedPersonIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: uuid
    UpdateMemoryRequest:
      type: object
      minProperties: 1
      description: Fields irrelevant to the existing memory type are rejected.
      properties:
        title:
          type: [string, 'null']
          maxLength: 250
        content:
          type: [string, 'null']
          maxLength: 50000
        caption:
          type: [string, 'null']
          maxLength: 5000
        takenAt:
          oneOf:
            - $ref: '#/components/schemas/PartialDate'
            - type: 'null'
        relatedPersonIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: uuid
    MemoryPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MemoryResponse'
        page:
          $ref: '#/components/schemas/PageMeta'
    MediaPurpose:
      type: string
      enum: [PROFILE_PICTURE, MEMORY_PHOTO]
    MediaStatus:
      type: string
      enum: [PENDING_UPLOAD, READY, FAILED, ARCHIVED]
...
    MediaAssetResponse:
      type: object
      required: [id, purpose, status, mimeType]
      properties:
        id:
          type: string
          format: uuid
        purpose:
          $ref: '#/components/schemas/MediaPurpose'
        status:
          $ref: '#/components/schemas/MediaStatus'
        mimeType:
          type: string
          description: MIME type declared at upload. The served derivatives are always JPEG (OQ-045).
        sizeBytes:
          type: [integer, 'null']
          format: int64
          description: Size declared at upload, in bytes (OQ-045).
        widthPx:
          type: [integer, 'null']
          description: Width of the display derivative. Present when READY.
        heightPx:
          type: [integer, 'null']
          description: Height of the display derivative. Present when READY.
        url:
          type: [string, 'null']
          format: uri
          description: Display derivative (long edge ≤ 2048 px). Pre-signed, valid 60 minutes. Present when READY.
        thumbnailUrl:
          type: [string, 'null']
          format: uri
          description: Thumbnail derivative (long edge ≤ 480 px). Pre-signed, valid 60 minutes. Present when READY.
    ActivityActor:
      type: object
      required: [userId, displayName, deleted]
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: [string, 'null']
          description: Null when the account was deleted; the UI then shows "Former member".
        deleted:
          type: boolean
...
    ProblemDetails:
      type: object
      required: [type, title, status, code, detail, traceId]
      properties:
        type:
          type: string
          format: uri
          example: https://mbia.example.com/problems/concurrent-modification
        title:
          type: string
          example: Concurrent modification
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          description: Stable machine-readable Mbia error code.
          examples:
            - CONCURRENT_MODIFICATION
            - PERSON_NOT_FOUND
            - MEMBERSHIP_REQUIRED
            - PERMISSION_DENIED
            - POSSIBLE_DUPLICATE
            - PERSON_ALREADY_CLAIMED
            - USER_ALREADY_LINKED
            - RELATIONSHIP_ALREADY_EXISTS
            - SELF_RELATIONSHIP_NOT_ALLOWED
            - RELATIONSHIP_CREATES_CYCLE
            - RELATIONSHIP_WARNING_CONFIRMATION_REQUIRED
            - PERSON_NOT_ACTIVE
            - PERSON_MERGE_CONFLICT
            - EMAIL_NOT_VERIFIED
            - LAST_ADMIN_REQUIRED
            - INVITATION_NOT_FOUND
            - INVITATION_EXPIRED
            - INVITATION_REVOKED
            - INVITATION_ALREADY_USED
            - MEDIA_TOO_LARGE
            - MEDIA_INVALID
            - MEDIA_NOT_READY
            - MEDIA_ALREADY_USED
            - MEDIA_NOT_FOUND
            - MEMORY_NOT_FOUND
            - VALIDATION_FAILED
            - INTERNAL_ERROR
            - AUTHENTICATION_REQUIRED
            - FAMILY_NOT_FOUND
            - RESOURCE_NOT_FOUND
            - METHOD_NOT_ALLOWED
            - NOT_ACCEPTABLE
            - UNSUPPORTED_MEDIA_TYPE
        detail:
          type: string
        traceId:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        details:
          type: object
          additionalProperties: true
          description: Structured domain-specific context, e.g. duplicate candidates or relationship warnings.
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string